Switch to google places search
