+++ hooks/useDirector.ts
@@ -4,9 +4,9 @@ import Director from "../lib/director"
 import Router from "../lib/router"
 import { Vibration } from "react-native"
 
-const useDirector = ({ vibrate = true }: { vibrate: boolean }) => {
+const useDirector = ({ vibrate = true, apiKey }: { vibrate: boolean, apiKey: string }) => {
   const result = useMemo(() => {
-    const router = new Router(process.env.EXPO_PUBLIC_MAPBOX_TOKEN as string);
+    const router = new Router(apiKey);
     const director = new Director(router);
     const vibration = vibrate ? Vibration.vibrate : () => { }
     const vibrator = new Vibrator(director, vibration);
