+++ package.json
@@ -13,7 +13,6 @@
   },
   "dependencies": {
     "@mapbox/mapbox-sdk": "^0.15.3",
-    "@mapbox/search-js-core": "^1.0.0-beta.18",
     "@turf/turf": "^6.5.0",
     "@types/lodash": "^4.14.202",
     "events": "^3.3.0",
@@ -22,7 +21,6 @@
     "expo-location": "~16.1.0",
     "expo-status-bar": "~1.6.0",
     "expo-updates": "~0.18.19",
-    "lodash": "^4.17.21",
     "react": "18.2.0",
     "react-native": "0.72.6",
     "react-native-google-places-autocomplete": "^2.5.6",
